requirements.txt
accelerate>=0.16.0
torchvision
transformers>=4.25.1
datasets
ftfy